Refactor blog structure and remove unused files
- Commented out blog post imports and related URL generation in `sitemap.ts` to streamline sitemap creation.
- Deleted unused blog-related files, including authors, blog posts, and manifest, to clean up the codebase.
- Updated `worker.tsx` to import blog routes from a new location, ensuring proper routing functionality.
- Removed the `BlogList`, `BlogPage`, and `Post` components to eliminate redundancy and improve maintainability.
