+++ src/app/addons/blog/data/posts/full-stack-colocation.md
