+++ src/app/addons/blog/data/posts/full-stack-payment-integration.md
