+++ src/app/addons/blog/index.ts
@@ -1,0 +1,3 @@
+import { blogRoutes } from "./routes";
+
+export { blogRoutes };
