--- src/app/pages/blog/layout.tsx
@@ -1,1 +1,0 @@
-
